requirements.txt
argcomplete==1.9.3
boto3==1.9.148
botocore==1.12.148
certifi==2019.3.9
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-filter==2.1.0
django-rest-swagger==2.2.0
django-s3-storage==0.12.4
django-storages==1.7.1
djangorestframework==3.9.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
Markdown==3.1
MarkupSafe==1.1.1
numpy==1.16.3
openapi-codec==1.3.2
pandas==0.24.2
placebo==0.9.0
psycopg2-binary==2.7.7
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.9
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.6
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.15.2
wsgi-request-logger==0.4.6
xlrd==1.2.0
zappa==0.48.2
